# server/docs/swagger.yaml
basePath: /api
definitions:
  github_com_yogesh4952_ebookstore_internal_address_models.UserAddress:
    properties:
      city:
        type: string
      createdAt:
        type: string
      deletedAt:
        $ref: '#/definitions/gorm.DeletedAt'
      delivery_address:
        type: string
      id:
        type: integer
      updatedAt:
        type: string
      user:
        $ref: '#/definitions/github_com_yogesh4952_ebookstore_internal_user_models.User'
      user_id:
        type: integer
    required:
    - city
    - delivery_address
    - user_id
    type: object
  github_com_yogesh4952_ebookstore_internal_auth_models.RegisterPayload:
    properties:
      age:
        type: integer
      email:
        type: string
      first_name:
        type: string
      last_name:
        type: string
      role:
        $ref: '#/definitions/github_com_yogesh4952_ebookstore_internal_user_models.Role'
    required:
    - age
    - email
    - first_name
    - last_name
    - role
    type: object
  github_com_yogesh4952_ebookstore_internal_book_models.PublishBookPayload:
    properties:
      author_name:
        type: string
      category:
        type: string
      cover_page_url:
        type: string
      genre:
        type: string
      pages:
        minimum: 10
        type: integer
      price:
        minimum: 10
        type: number
      publication:
        type: string
      seller_id:
        type: integer
      title:
        maxLength: 60
        minLength: 2
        type: string
      units:
        minimum: 0
        type: integer
    required:
    - author_name
    - category
    - genre
    - pages
    - price
    - publication
    - title
    - units
    type: object
  github_com_yogesh4952_ebookstore_internal_book_models.UpdateBookPayload:
    properties:
      author_name:
        type: string
      book_id:
        type: integer
      category:
        type: string
      cover_page_url:
        type: string
      genre:
        type: string
      pages:
        minimum: 10
        type: integer
      price:
        minimum: 10
        type: number
      publication:
        type: string
      seller_id:
        type: integer
      title:
        maxLength: 60
        minLength: 2
        type: string
      units:
        minimum: 0
        type: integer
    required:
    - book_id
    - seller_id
    type: object
...
  github_com_yogesh4952_ebookstore_internal_user_models.Role:
    enum:
    - admin
    - customer
    - seller
    type: string
    x-enum-varnames:
    - RoleAdmin
    - RoleCustomer
    - Roleseller
  github_com_yogesh4952_ebookstore_internal_user_models.User:
    properties:
      createdAt:
        type: string
      deletedAt:
        $ref: '#/definitions/gorm.DeletedAt'
      email:
        type: string
      first_name:
        type: string
      id:
        type: integer
      last_name:
        type: string
      phone_number:
        type: string
      role:
        $ref: '#/definitions/github_com_yogesh4952_ebookstore_internal_user_models.Role'
      updatedAt:
        type: string
    type: object
  gorm.DeletedAt:
    properties:
      time:
        type: string
      valid:
        description: Valid is true if Time is not NULL
        type: boolean
    type: object
  internal_auth_handler.LoginPayload:
    properties:
      email:
        type: string
      otp:
        type: string
    required:
    - email
    - otp
    type: object
  internal_auth_handler.SendOTPRequest:
    properties:
      email:
        type: string
    required:
    - email
    type: object
host: localhost:8080
info:
  contact: {}
  description: REST API for the EBook Store platform
  title: EBook Store API
  version: "1.0"
paths:
  /address/add-address:
    post:
      consumes:
      - application/json
      description: Add a shipping address. The `user_id` field is taken from the request
        body and is not checked against the authenticated user, so any authenticated
        caller can currently attach an address to another user's account. Both the
        city and the delivery address are required.
      parameters:
      - description: Address data
        in: body
        name: payload
        required: true
        schema:
          $ref: '#/definitions/github_com_yogesh4952_ebookstore_internal_address_models.UserAddress'
      produces:
      - application/json
      responses:
        "202":
          description: Address added, containing success and message keys
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Invalid request body, the user does not exist, or the address
            could not be saved
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Missing or invalid token
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Add a new address
      tags:
      - addresses
  /auth/login:
    post:
      consumes:
      - application/json
      description: Authenticate a user with the email and the one-time code sent by
        `/auth/send-otp`, and return a JWT access token. The OTP is consumed on success
        and cannot be reused. Note that the success response contains `message` and
        `token` keys only, with no `success` key, whereas the error responses use
        an `error` key.
      parameters:
      - description: Email and OTP
        in: body
        name: payload
        required: true
        schema:
          $ref: '#/definitions/internal_auth_handler.LoginPayload'
      produces:
      - application/json
      responses:
        "200":
          description: Login successful, containing message and token keys
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Invalid request body, containing an error key
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Incorrect or expired OTP, containing an error key
          schema:
            additionalProperties: true
            type: object
      summary: Login with email and OTP
      tags:
      - auth
  /auth/register:
    post:
      consumes:
      - application/json
      description: Create a new user account as either a customer or a seller. Registering
        with an email that already exists returns 409. Use a role of `admin` only
        for internal tooling; normal signups should use `customer` or `seller`.
      parameters:
      - description: Registration data
        in: body
        name: payload
        required: true
        schema:
          $ref: '#/definitions/github_com_yogesh4952_ebookstore_internal_auth_models.RegisterPayload'
      produces:
      - application/json
      responses:
        "200":
          description: Account created, containing success and message keys
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Invalid request body, or an invalid role
          schema:
            additionalProperties: true
            type: object
        "409":
          description: Email is already registered
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Account could not be created
          schema:
            additionalProperties: true
            type: object
      summary: Register a new user
      tags:
      - auth
  /auth/send-otp:
    post:
      consumes:
      - application/json
      description: 'Send a one-time password to the user''s email for verification.
        The code is valid for five minutes and can only be used once. Note that this
        endpoint uses two different response envelopes: a 400 returns `{"error": "..."}`
        while a 500 returns `{"success": false, "message": "..."}`.'
      parameters:
      - description: Email address
        in: body
        name: payload
        required: true
        schema:
          $ref: '#/definitions/internal_auth_handler.SendOTPRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OTP sent, containing success and message keys
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Invalid request body, containing an error key
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Unknown email, or the code could not be generated, stored,
            or emailed
          schema:
            additionalProperties: true
            type: object
      summary: Send OTP to email
      tags:
      - auth
  /book/list-books:
    get:
      description: Get a paginated list of books. The response echoes the `page` and
        `limit` actually applied by the server, which may differ from the requested
        values when out of range. Any service or database error is reported as 404.
      parameters:
      - default: 1
        description: Page number
        in: query
        name: page
        type: integer
      - default: 10
        description: Items per page
        in: query
        name: limit
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Books fetched, containing success, message, the books under
            data, and the total, page, and limit counts
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Books could not be fetched
          schema:
            additionalProperties: true
            type: object
      summary: List all books
      tags:
      - books
  /book/publish-book:
    post:
      consumes:
      - application/json
      description: Create a new book listing. Restricted to the `seller` and `admin`
        roles; any other authenticated user receives 403. The authenticated user id
        is taken from the JWT and used as the book owner, so it cannot be spoofed
        in the request body.
      parameters:
      - description: Book data
        in: body
        name: payload
        required: true
        schema:
          $ref: '#/definitions/github_com_yogesh4952_ebookstore_internal_book_models.PublishBookPayload'
      produces:
      - application/json
      responses:
        "201":
          description: Book published, containing success and message keys
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Invalid request body, or the book could not be saved
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Missing or invalid token, or no user identity on the request
            context
          schema:
            additionalProperties: true
            type: object
        "403":
          description: Authenticated user is not a seller or admin
          schema:
            additionalProperties: true
            type: object
        "500":
          description: User id on the request context had an unexpected type
          schema:
            additionalProperties: true
            type: object
        "502":
          description: The book service failed to process the request
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Publish a new book
      tags:
      - books
  /book/update-book:
    patch:
      consumes:
      - application/json
      description: |-
        Update book details (owner only)
        Update the details of a book the authenticated user owns. Only the fields present in the request body are modified. Restricted to the `seller` and `admin` roles; any other authenticated user receives 403.
      parameters:
      - description: Updated book data
        in: body
        name: payload
        required: true
        schema:
          $ref: '#/definitions/github_com_yogesh4952_ebookstore_internal_book_models.UpdateBookPayload'
      produces:
      - application/json
      responses:
        "202":
          description: Book updated, containing success, message, and the updated
            book under data
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Missing user identity, invalid request body, or the update
            failed
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Missing or invalid token
          schema:
            additionalProperties: true
            type: object
        "403":
          description: Authenticated user is not a seller or admin
          schema:
            additionalProperties: true
            type: object
        "500":
          description: User id on the request context had an unexpected type
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Update an existing book
      tags:
      - books
...
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"